Conduct K Means Clustering

Conduct K Means Clustering

Usage

k_means(min_k, max_k, n_starts = 25, seed = 42, pca)

Arguments

min_k

The min amount k value you wish to start at

max_k

The max amount k value you wish to go up to

n_starts

Number of initial sets

seed

Random seed for reproducibility

pca

Resulting PCA matrix

Value

List Average silhouette scores and of K Means clustering results with various K values

Examples

library(analysiskmeans)
data(example_sce)
SingleCellExperiment::logcounts(example_sce) <- log1p(SingleCellExperiment::counts(example_sce))#Line taken from Claude AI - Debug
sce <- example_sce
results <- data_config(sce)
#> Warning: 'librarySizeFactors' is deprecated.
#> Use 'scrapper::centerSizeFactors' instead.
#> See help("Deprecated")
#> Warning: 'normalizeCounts' is deprecated.
#> Use 'scrapper::normalizeCounts' instead.
#> See help("Deprecated")
mat_norm <- top_x_genes(results$sce, n_top = 50)
pca <- computepca(mat_norm)
k_means(4, 9, n_starts = 25, seed = 42, pca)
#> $metrics
#>   k      wss avg_silhouette
#> 1 4 689.9560     0.05556822
#> 2 5 625.5026     0.06131297
#> 3 6 565.5828     0.06596026
#> 4 7 512.3719     0.06487671
#> 5 8 459.8210     0.06096352
#> 6 9 408.9954     0.06596343
#> 
#> $km_list
#> $km_list$`4`
#> K-means clustering with 4 clusters of sizes 5, 6, 4, 5
#> 
#> Cluster means:
#>          PC1        PC2         PC3        PC4        PC5        PC6
#> 1  0.8232067 -0.8839124  0.04523608  2.6112471 -0.9545080 -0.1528163
#> 2  1.6886721  2.5539593 -0.06304373 -1.0985761 -0.5922869  0.1821593
#> 3  0.6230835 -3.3298738 -0.59296210 -1.1663800  0.4735455  0.6938414
#> 4 -3.3480799  0.4830603  0.50478608 -0.3598518  1.2864159 -0.6208480
#>             PC7         PC8         PC9       PC10        PC11        PC12
#> 1 -0.1888941132  0.76621593  0.04956604  0.2115829 -0.24864958 -0.33774825
#> 2  0.0004984321  0.18664874 -0.15596771 -0.2554676  0.02378879  0.09522327
#> 3 -0.0821085803 -1.16140895 -0.30747673 -0.7436367  0.30815670  0.37444744
#> 4  0.2539828590 -0.06106726  0.38357659  0.6898875 -0.02642233 -0.07607763
#>         PC13        PC14       PC15         PC16         PC17        PC18
#> 1 -0.2801969 -0.30766032 -0.1919007  0.471046708 -0.105349781  0.04905483
#> 2  0.2591522  0.26934504  0.1066981 -0.197540193  0.032224804 -0.16374325
#> 3 -0.3959851  0.10326587  0.2675799 -0.282831961  0.094656634 -0.07383421
#> 4  0.2860023 -0.09816642 -0.1502010 -0.007732908 -0.009045291  0.20650444
#>          PC19         PC20
#> 1  0.03405607 4.773770e-16
#> 2  0.04369059 3.139096e-16
#> 3 -0.15219664 2.383870e-16
#> 4  0.03527252 5.136058e-16
#> 
#> Clustering vector:
#>  [1] 2 2 1 1 4 4 2 2 3 4 3 4 2 4 1 1 3 3 1 2
#> 
#> Within cluster sum of squares by cluster:
#> [1] 166.1667 219.2782 113.2901 191.2211
#>  (between_SS / total_SS =  27.4 %)
#> 
#> Available components:
#> 
#> [1] "cluster"      "centers"      "totss"        "withinss"     "tot.withinss"
#> [6] "betweenss"    "size"         "iter"         "ifault"      
#> 
#> $km_list$`5`
#> K-means clustering with 5 clusters of sizes 6, 1, 4, 5, 4
#> 
#> Cluster means:
#>          PC1         PC2         PC3        PC4         PC5        PC6
#> 1  1.6886721  2.55395926 -0.06304373 -1.0985761 -0.59228688  0.1821593
#> 2 -1.8250694  2.11683800 -0.09109240  1.8888415  6.11579878  2.2341937
#> 3 -3.7288325  0.07461588  0.65375570 -0.9220252  0.07907016 -1.3346084
#> 4  0.8232067 -0.88391237  0.04523608  2.6112471 -0.95450799 -0.1528163
#> 5  0.6230835 -3.32987382 -0.59296210 -1.1663800  0.47354545  0.6938414
#>             PC7          PC8         PC9       PC10        PC11         PC12
#> 1  0.0004984321  0.186648744 -0.15596771 -0.2554676  0.02378879  0.095223266
#> 2 -1.5523365244 -0.272478348  1.50412075 -0.8662418  0.61202909 -0.414401426
#> 3  0.7055627049 -0.008214494  0.10344055  1.0789199 -0.18603519  0.008503324
#> 4 -0.1888941132  0.766215932  0.04956604  0.2115829 -0.24864958 -0.337748247
#> 5 -0.0821085803 -1.161408951 -0.30747673 -0.7436367  0.30815670  0.374447442
#>          PC13        PC14        PC15        PC16        PC17        PC18
#> 1  0.25915224  0.26934504  0.10669815 -0.19754019  0.03222480 -0.16374325
#> 2  1.07070939 -0.84692480 -0.47097969 -0.25515116 -0.90119103  0.19145499
#> 3  0.08982547  0.08902317 -0.07000629  0.05412166  0.21399114  0.21026680
#> 4 -0.28019688 -0.30766032 -0.19190072  0.47104671 -0.10534978  0.04905483
#> 5 -0.39598508  0.10326587  0.26757989 -0.28283196  0.09465663 -0.07383421
#>          PC19         PC20
#> 1  0.04369059 3.139096e-16
#> 2 -0.17610461 7.168004e-16
#> 3  0.08811681 4.628072e-16
#> 4  0.03405607 4.773770e-16
#> 5 -0.15219664 2.383870e-16
#> 
#> Clustering vector:
#>  [1] 1 1 4 4 3 2 1 1 5 3 5 3 1 3 4 4 5 5 4 1
#> 
#> Within cluster sum of squares by cluster:
#> [1] 219.2782   0.0000 126.7676 166.1667 113.2901
#>  (between_SS / total_SS =  34.2 %)
#> 
#> Available components:
#> 
#> [1] "cluster"      "centers"      "totss"        "withinss"     "tot.withinss"
#> [6] "betweenss"    "size"         "iter"         "ifault"      
#> 
#> $km_list$`6`
#> K-means clustering with 6 clusters of sizes 4, 1, 4, 2, 5, 4
#> 
#> Cluster means:
#>          PC1         PC2        PC3        PC4         PC5         PC6
#> 1  1.8594103  3.10127579  1.2870529 -0.8516263 -0.88034246  0.49950347
#> 2 -1.8250694  2.11683800 -0.0910924  1.8888415  6.11579878  2.23419374
#> 3  0.6230835 -3.32987382 -0.5929621 -1.1663800  0.47354545  0.69384138
#> 4  0.8666903 -1.26000553  2.6782478  2.8660732 -0.67259069 -0.87710064
#> 5  1.0154088  0.20382033 -2.1313579  0.8278276 -0.69194201  0.01701237
#> 6 -3.7288325  0.07461588  0.6537557 -0.9220252  0.07907016 -1.33460842
#>           PC7          PC8        PC9        PC10       PC11         PC12
#> 1  0.05496919 -1.080437934 -0.4007576  0.02909602  0.4479306 -0.131171716
#> 2 -1.55233652 -0.272478348  1.5041207 -0.86624176  0.6120291 -0.414401426
#> 3 -0.08210858 -1.161408951 -0.3074767 -0.74363668  0.3081567  0.374447442
#> 4 -2.20268516  1.012357474 -0.7154252  0.16657577 -0.4056814 -1.096222379
#> 5  0.64880271  1.449601782  0.4691809 -0.18488531 -0.4161750  0.319945997
#> 6  0.70556270 -0.008214494  0.1034405  1.07891985 -0.1860352  0.008503324
#>          PC13        PC14        PC15        PC16         PC17        PC18
#> 1  0.17750003  0.04031245  0.36205828  0.15238809  0.034039902 -0.03707475
#> 2  1.07070939 -0.84692480 -0.47097969 -0.25515116 -0.901191028  0.19145499
#> 3 -0.39598508  0.10326587  0.26757989 -0.28283196  0.094656634 -0.07383421
#> 4 -0.67899342  0.67965884 -0.80355681 -0.40716320 -0.009051229 -0.19439741
#> 5  0.16038315 -0.28855977 -0.03208684  0.27495328 -0.090291446 -0.04001830
#> 6  0.08982547  0.08902317 -0.07000629  0.05412166  0.213991144  0.21026680
#>          PC19         PC20
#> 1 -0.10637791 2.294383e-16
#> 2 -0.17610461 7.168004e-16
#> 3 -0.15219664 2.383870e-16
#> 4 -0.15386157 5.926884e-16
#> 5  0.23313174 4.334425e-16
#> 6  0.08811681 4.628072e-16
#> 
#> Clustering vector:
#>  [1] 5 5 4 5 6 2 1 1 3 6 3 6 1 6 5 4 3 3 5 1
#> 
#> Within cluster sum of squares by cluster:
#> [1] 117.14363   0.00000 113.29009  33.56533 174.81615 126.76763
#>  (between_SS / total_SS =  40.5 %)
#> 
#> Available components:
#> 
#> [1] "cluster"      "centers"      "totss"        "withinss"     "tot.withinss"
#> [6] "betweenss"    "size"         "iter"         "ifault"      
#> 
#> $km_list$`7`
#> K-means clustering with 7 clusters of sizes 4, 3, 1, 2, 4, 4, 2
#> 
#> Cluster means:
#>          PC1         PC2        PC3        PC4         PC5        PC6
#> 1  0.6230835 -3.32987382 -0.5929621 -1.1663800  0.47354545  0.6938414
#> 2  0.5366099  0.02493894 -2.7589657  0.4801255 -1.56670705  1.3305803
#> 3 -1.8250694  2.11683800 -0.0910924  1.8888415  6.11579878  2.2341937
#> 4  0.8666903 -1.26000553  2.6782478  2.8660732 -0.67259069 -0.8771006
#> 5 -3.7288325  0.07461588  0.6537557 -0.9220252  0.07907016 -1.3346084
#> 6  1.8594103  3.10127579  1.2870529 -0.8516263 -0.88034246  0.4995035
#> 7  1.7336071  0.47214242 -1.1899461  1.3493807  0.62020554 -1.9533396
#>           PC7          PC8        PC9        PC10       PC11         PC12
#> 1 -0.08210858 -1.161408951 -0.3074767 -0.74363668  0.3081567  0.374447442
#> 2 -0.45668326  1.440384037  1.5215258  0.69363263 -0.4846333 -0.346429176
#> 3 -1.55233652 -0.272478348  1.5041207 -0.86624176  0.6120291 -0.414401426
#> 4 -2.20268516  1.012357474 -0.7154252  0.16657577 -0.4056814 -1.096222379
#> 5  0.70556270 -0.008214494  0.1034405  1.07891985 -0.1860352  0.008503324
#> 6  0.05496919 -1.080437934 -0.4007576  0.02909602  0.4479306 -0.131171716
#> 7  2.30703168  1.463428401 -1.1093364 -1.50266222 -0.3134875  1.319508755
#>          PC13        PC14        PC15        PC16         PC17        PC18
#> 1 -0.39598508  0.10326587  0.26757989 -0.28283196  0.094656634 -0.07383421
#> 2 -0.02960579 -0.51750273  0.15341192  0.37220440  0.229148007  0.02592810
#> 3  1.07070939 -0.84692480 -0.47097969 -0.25515116 -0.901191028  0.19145499
#> 4 -0.67899342  0.67965884 -0.80355681 -0.40716320 -0.009051229 -0.19439741
#> 5  0.08982547  0.08902317 -0.07000629  0.05412166  0.213991144  0.21026680
#> 6  0.17750003  0.04031245  0.36205828  0.15238809  0.034039902 -0.03707475
#> 7  0.44536656  0.05485467 -0.31033498  0.12907661 -0.569450625 -0.13893790
#>          PC19         PC20
#> 1 -0.15219664 2.383870e-16
#> 2  0.03523431 4.118141e-16
#> 3 -0.17610461 7.168004e-16
#> 4 -0.15386157 5.926884e-16
#> 5  0.08811681 4.628072e-16
#> 6 -0.10637791 2.294383e-16
#> 7  0.52997789 4.658852e-16
#> 
#> Clustering vector:
#>  [1] 2 7 4 7 5 3 6 6 1 5 1 5 6 5 2 4 1 1 2 6
#> 
#> Within cluster sum of squares by cluster:
#> [1] 113.29009  83.53243   0.00000  33.56533 126.76763 117.14363  38.07278
#>  (between_SS / total_SS =  46.1 %)
#> 
#> Available components:
#> 
#> [1] "cluster"      "centers"      "totss"        "withinss"     "tot.withinss"
#> [6] "betweenss"    "size"         "iter"         "ifault"      
#> 
#> $km_list$`8`
#> K-means clustering with 8 clusters of sizes 3, 3, 2, 3, 2, 3, 3, 1
#> 
#> Cluster means:
#>          PC1        PC2        PC3        PC4         PC5        PC6
#> 1 -2.6431756  2.6439260 -1.5618124 -0.5785339 -1.43679216  0.7123916
#> 2  0.9341642 -1.4481344 -1.9794286  1.6136774 -0.75415324  1.8148803
#> 3  0.8666903 -1.2600055  2.6782478  2.8660732 -0.67259069 -0.8771006
#> 4  2.9106736  2.6490276  1.9783790 -1.5102635 -0.76416896  0.4256750
#> 5 -2.6421492  0.3480806 -0.5058378 -1.1920905  0.37957443 -3.3619387
#> 6 -1.1748566 -3.3143947  1.3966319 -2.1379445 -0.01571026  1.2228171
#> 7  1.7651901 -0.6280872 -1.2516790  0.8674622  1.12756919 -2.0944690
#> 8 -1.8250694  2.1168380 -0.0910924  1.8888415  6.11579878  2.2341937
#>           PC7        PC8        PC9       PC10         PC11       PC12
#> 1 -0.40315332  0.5992359 -0.5786167 -1.1029554 -0.366531168 -0.1784113
#> 2  0.20580556 -0.1670225  0.4774698  1.6985025  0.026734624  0.9427638
#> 3 -2.20268516  1.0123575 -0.7154252  0.1665758 -0.405681362 -1.0962224
#> 4 -0.03270814 -1.0955624 -0.1421632  0.4306978  0.165906286 -0.1857134
#> 5  0.19813922 -0.8201552  1.3848421  1.9966762  0.545829462  0.1778028
#> 6  0.74674454  0.1998996 -0.2278211 -0.5053682 -0.125196668 -0.3470103
#> 7  1.33712082  0.4261407 -0.4765202 -1.6742974  0.001645163  0.5187846
#> 8 -1.55233652 -0.2724783  1.5041207 -0.8662418  0.612029093 -0.4144014
#>         PC13         PC14        PC15        PC16         PC17         PC18
#> 1 -0.2957578  0.315101998  0.26626947 -0.33448882  0.427102497 -0.015192054
#> 2 -0.1200446 -0.550075367 -0.06302744  0.26726909  0.452727666  0.031088612
#> 3 -0.6789934  0.679658844 -0.80355681 -0.40716320 -0.009051229 -0.194397414
#> 4  0.2385064 -0.470120349  0.09915466 -0.05635828 -0.095468377  0.434042904
#> 5 -0.7212556  0.373534939 -0.37714550  0.20222131 -0.547518861 -0.033357108
#> 6  0.6516392  0.280618511  0.21552965  0.41517207 -0.406483761 -0.352646581
#> 7  0.1022530  0.004654284  0.42620176 -0.06991574  0.293565711 -0.009274864
#> 8  1.0707094 -0.846924801 -0.47097969 -0.25515116 -0.901191028  0.191454990
#>          PC19         PC20
#> 1 -0.41965643 5.833698e-16
#> 2 -0.10208602 9.895580e-17
#> 3 -0.15386157 5.926884e-16
#> 4  0.32253916 7.964936e-17
#> 5 -0.05313207 4.029041e-16
#> 6 -0.07979088 1.594868e-16
#> 7  0.47569147 7.731830e-16
#> 8 -0.17610461 7.168004e-16
#> 
#> Clustering vector:
#>  [1] 1 7 3 7 1 8 4 1 6 6 6 5 4 5 2 3 7 2 2 4
#> 
#> Within cluster sum of squares by cluster:
#> [1] 80.63448 72.03770 33.56533 76.84044 39.77393 80.62319 76.34593  0.00000
#>  (between_SS / total_SS =  51.6 %)
#> 
#> Available components:
#> 
#> [1] "cluster"      "centers"      "totss"        "withinss"     "tot.withinss"
#> [6] "betweenss"    "size"         "iter"         "ifault"      
#> 
#> $km_list$`9`
#> K-means clustering with 9 clusters of sizes 3, 2, 3, 2, 1, 2, 2, 2, 3
#> 
#> Cluster means:
#>          PC1        PC2        PC3        PC4         PC5        PC6
#> 1  2.9106736  2.6490276  1.9783790 -1.5102635 -0.76416896  0.4256750
#> 2  0.8666903 -1.2600055  2.6782478  2.8660732 -0.67259069 -0.8771006
#> 3  0.9341642 -1.4481344 -1.9794286  1.6136774 -0.75415324  1.8148803
#> 4 -2.0056567 -2.8595359  2.6768906 -1.2594449  0.03351467  1.7199088
#> 5 -1.8250694  2.1168380 -0.0910924  1.8888415  6.11579878  2.2341937
#> 6  1.1575499 -3.5263293 -1.2695153 -1.9956592  1.01406818 -1.0740471
#> 7 -2.6421492  0.3480806 -0.5058378 -1.1920905  0.37957443 -3.3619387
#> 8  1.7336071  0.4721424 -1.1899461  1.3493807  0.62020554 -1.9533396
#> 9 -2.6431756  2.6439260 -1.5618124 -0.5785339 -1.43679216  0.7123916
#>           PC7        PC8        PC9        PC10        PC11       PC12
#> 1 -0.03270814 -1.0955624 -0.1421632  0.43069785  0.16590629 -0.1857134
#> 2 -2.20268516  1.0123575 -0.7154252  0.16657577 -0.40568136 -1.0962224
#> 3  0.20580556 -0.1670225  0.4774698  1.69850245  0.02673462  0.9427638
#> 4  1.75856396  0.7944549  0.4879130  0.08107176 -0.29998273 -0.1835439
#> 5 -1.55233652 -0.2724783  1.5041207 -0.86624176  0.61202909 -0.4144014
#> 6 -0.93979760 -1.3188228 -0.4350886 -1.84790807  0.42814297 -0.8783034
#> 7  0.19813922 -0.8201552  1.3848421  1.99667623  0.54582946  0.1778028
#> 8  2.30703168  1.4634284 -1.1093364 -1.50266222 -0.31348750  1.3195088
#> 9 -0.40315332  0.5992359 -0.5786167 -1.10295537 -0.36653117 -0.1784113
#>         PC13        PC14        PC15        PC16         PC17        PC18
#> 1  0.2385064 -0.47012035  0.09915466 -0.05635828 -0.095468377  0.43404290
#> 2 -0.6789934  0.67965884 -0.80355681 -0.40716320 -0.009051229 -0.19439741
#> 3 -0.1200446 -0.55007537 -0.06302744  0.26726909  0.452727666  0.03108861
#> 4  0.4934948  0.70166001  1.22315664 -0.16815240 -0.449924850 -0.36550261
#> 5  1.0707094 -0.84692480 -0.47097969 -0.25515116 -0.901191028  0.19145499
#> 6  0.1919770 -0.32860549  0.04977545  0.55696028  0.849998400 -0.03844166
#> 7 -0.7212556  0.37353494 -0.37714550  0.20222131 -0.547518861 -0.03335711
#> 8  0.4453666  0.05485467 -0.31033498  0.12907661 -0.569450625 -0.13893790
#> 9 -0.2957578  0.31510200  0.26626947 -0.33448882  0.427102497 -0.01519205
#>          PC19         PC20
#> 1  0.32253916 7.964936e-17
#> 2 -0.15386157 5.926884e-16
#> 3 -0.10208602 9.895580e-17
#> 4  0.21010757 1.556914e-16
#> 5 -0.17610461 7.168004e-16
#> 6 -0.14623458 7.774282e-16
#> 7 -0.05313207 4.029041e-16
#> 8  0.52997789 4.658852e-16
#> 9 -0.41965643 5.833698e-16
#> 
#> Clustering vector:
#>  [1] 9 8 2 8 9 5 1 9 4 4 6 7 1 7 3 2 6 3 3 1
#> 
#> Within cluster sum of squares by cluster:
#> [1] 76.84044 33.56533 72.03770 35.67020  0.00000 32.40049 39.77393 38.07278
#> [9] 80.63448
#>  (between_SS / total_SS =  56.9 %)
#> 
#> Available components:
#> 
#> [1] "cluster"      "centers"      "totss"        "withinss"     "tot.withinss"
#> [6] "betweenss"    "size"         "iter"         "ifault"      
#> 
#>